+++ packages/headless/src/features/facets/automatic-facet-set/interfaces/request.ts
@@ -2,5 +2,6 @@ import {FacetValue} from '../../../../product-listing.index';
 
 export type AutomaticFacetRequest = {
   field: string;
+  label: string;
   currentValues: FacetValue[];
 };